Done: - Table Visualization Engine that shows the logs in a presentable manner. - Organized the whole csproj more To Do: - ReadMe.md file
Due Date: 12/30/2024

+++ Controllers/DatabaseController.cs
@@ -1,15 +1,19 @@
-﻿using Dapper;
+﻿using CodingTracker.Engines;
+using CodingTracker.Models;
+using Dapper;
 using Spectre.Console;
 using System.Configuration;
 using System.Data.SQLite;
 
-namespace CodingTracker
+namespace CodingTracker.Controllers
 {
-    public class CodingLogsDatabase
+    public class DatabaseController
     {
+
         SQLiteConnection connection;
-        public CodingLogsDatabase()
+        public DatabaseController()
         {
+
             string? connnectionString = ConfigurationManager.AppSettings["ConnectionString"];
             // Create a new SQLite database
             connection = new SQLiteConnection(connnectionString);
@@ -32,34 +36,38 @@ namespace CodingTracker
         }
 
         //Delete a coding session
-        public void DeleteLog(int id)
+        public bool DeleteLog(int id)
         {
-            var sql = "DELETE FROM Logs WHERE Id = @id";
-            object[] param = { new { id = id } };
-            connection.Execute(sql, param);
+            bool result = false;
+            var sql = "SELECT COUNT(*) FROM Logs";
+            int count = connection.ExecuteScalar<int>(sql);
+            if (count > 0)
+            {
+                result = true;
+                sql = "DELETE FROM Logs WHERE Id = @id";
+                object[] param = { new { id } };
+                connection.Execute(sql, param);
 
-            sql = @"UPDATE Logs SET id = id - 1 WHERE id > @id";
-            connection.Execute(sql, param);
+                sql = @"UPDATE Logs SET id = id - 1 WHERE id > @id";
+                connection.Execute(sql, param);
+            }
+
+            return result;
         }
         //Review coding sessions
-        public void ReviewLogs()
+        public List<CodingSession> GetLogs()
         {
             AnsiConsole.Clear();
-
             var sql = "SELECT * FROM Logs";
             List<CodingSession> logs = (List<CodingSession>)connection.Query<CodingSession>(sql);
+            return logs;
+        }
 
-            foreach (CodingSession log in logs)
-            {
-                if (log.duration == 1)
-                {
-                    AnsiConsole.MarkupLine($"Id: {log.id}, Name: {log.name}, Start Time: {log.startTime}, End Time: {log.endTime}, Duration: {log.duration} hour");
-                }
-                else
-                {
-                    AnsiConsole.MarkupLine($"Id: {log.id}, Name: {log.name}, Start Time: {log.startTime}, End Time: {log.endTime}, Duration: {log.duration} hours");
-                }
-            }
+        public int GetLogCount()
+        {
+            var sql = "SELECT COUNT(*) FROM Logs";
+            int count = connection.ExecuteScalar<int>(sql);
+            return count;
         }
 
         //Quit the application
